requirements.txt
greenlet==3.1.1
numpy==2.2.0
pandas==2.2.3
pandas-stubs==2.2.3.241126
polars==1.17.1
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
types-psycopg2==2.9.21.20241019
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.1
